Split TUI presentation into components
